esp32-configurator/requirements.txt
argparse==1.4.0
jinja2==3.1.2
pydantic==1.10.2
pyyaml==6.0
